third_party/blink/web_tests/svg/dynamic-updates/SVGFEDisplacementMapElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS displacementMap.in1.baseVal is "Map"
PASS displacementMap.in1.baseVal is "Texture"
PASS successfullyParsed is true

TEST COMPLETE
